requirements.txt
tornado
pymongo
lxml
requests
pyyaml